Worked on more game settings
Added enough textures to test a full category.
Worked on implementing more logic to the difficulty system, still a WIP.
Fixed some bugs in the Game Manager that caused the game to end before finding all matching pairs.
Got rid of GameDataSO and moved the logic to GameSettings
Correctly and fully implemented the card controller logic.